pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = "Discord Python Hackpack"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.3.2"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pylint = "^3.1.0"
yapf = "^0.40.2"

[tool.poetry.scripts]
bot = "bot.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
